+++ datatig/jsonschemabuilder.py
@@ -22,13 +22,26 @@ def build_json_schema(fields: list, child_schema=False) -> BuildJSONSchemaResult
         key_bits = field.get_key().split("/")
         final_key = key_bits.pop()
         json_schema_insert = json_schema
+        field_schema, field_required = field.get_json_schema()
         for key_bit in key_bits:
+            if field_required:
+                json_schema_insert["required"] = (
+                    json_schema_insert["required"] + [final_key]
+                    if "required" in json_schema_insert
+                    else [key_bit]
+                )
             if not json_schema_insert["properties"].get(key_bit):
                 json_schema_insert["properties"][key_bit] = {
                     "type": "object",
                     "properties": {},
                 }
             json_schema_insert = json_schema_insert["properties"][key_bit]
-        json_schema_insert["properties"][final_key] = field.get_json_schema()
+        json_schema_insert["properties"][final_key] = field_schema
+        if field_required:
+            json_schema_insert["required"] = (
+                json_schema_insert["required"] + [final_key]
+                if "required" in json_schema_insert
+                else [final_key]
+            )
 
     return BuildJSONSchemaResults(json_schema)