+++ datatig/models/field_date.py
@@ -26,13 +26,13 @@ class FieldDateConfigModel(FieldConfigModel):
                     )
                 )
 
-    def get_json_schema(self) -> dict:
+    def get_json_schema(self) -> tuple[dict, bool]:
         return {
             "type": "string",
             "format": "date",
             "title": self._title,
             "description": self._description,
-        }
+        }, self._required
 
     def get_new_item_json(self):
         return None