+++ datatig/models/field_datetime.py
@@ -25,13 +25,13 @@ class FieldDateTimeConfigModel(FieldConfigModel):
                     )
                 )
 
-    def get_json_schema(self) -> dict:
+    def get_json_schema(self) -> tuple[dict, bool]:
         return {
             "type": "string",
             "format": "date-time",
             "title": self._title,
             "description": self._description,
-        }
+        }, self._required
 
     def get_new_item_json(self):
         return None