+++ datatig/models/field_url.py
@@ -6,13 +6,13 @@ class FieldURLConfigModel(FieldConfigModel):
     def get_type(self) -> str:
         return "url"
 
-    def get_json_schema(self) -> dict:
+    def get_json_schema(self) -> tuple[dict, bool]:
         return {
             "type": "string",
             "format": "uri",
             "title": self._title,
             "description": self._description,
-        }
+        }, self._required
 
     def get_new_item_json(self):
         return None